# docs/arch/API_Contract_v1.0.yaml
openapi: "3.0.3"
info:
  title: Learning OS API
  description: AI-independent learning execution system — REST API contract v1.0
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/projects/{pid}/tasks:
    get:
      summary: List all tasks in a project
      operationId: listTasks
      tags: [Tasks]
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, doing, done]
      responses:
        "200":
          description: Task list grouped by stage
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskSummary"
              example:
                tasks:
                  - id: "t1"
                    stage_id: "s1"
                    title: "变量与数据类型"
                    type: "theory"
                    xp: 20
                    status: "done"
                    sort_order: 1
                  - id: "t2"
                    stage_id: "s1"
                    title: "编写第一个计算器"
                    type: "practice"
                    xp: 30
                    status: "doing"
                    sort_order: 2
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
  /api/projects/{pid}/tasks/today:
    get:
      summary: Get today's available tasks
      description: |
        Returns tasks that are actionable today: either 'pending' or 'doing' status,
        with dependencies resolved. Tasks whose dependencies are not all 'done'
        are marked as 'blocked'.
      operationId: getTodayTasks
      tags: [Tasks]
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Today's actionable tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/TodayTask"
              example:
                date: "2026-06-08"
                tasks:
                  - id: "t2"
                    stage_id: "s1"
                    stage_title: "第一阶段"
                    title: "编写第一个计算器"
                    type: "practice"
                    xp: 20
                    status: "doing"
                    depends: []
                    blocked: false
                    check: "实现加减乘除四则运算"
                  - id: "t3"
                    stage_id: "s1"
                    stage_title: "第一阶段"
                    title: "学习笔记"
                    type: "output"
                    xp: 50
                    status: "pending"
                    depends: ["t2"]
                    blocked: true
                    check: "输出结构化笔记"
  /api/tasks/{tid}:
    get:
      summary: Get task detail
      operationId: getTask
      tags: [Tasks]
      parameters:
        - name: tid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Full task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
              example:
                id: "t1"
                stage_id: "s1"
                stage_title: "第一阶段：基础语法"
                project_id: "a1b2c3d4"
                title: "变量与数据类型"
                type: "theory"
                xp: 20
                estimate: 30
                status: "pending"
                depends: []
                check: "能说出 Python 四种基本数据类型的区别"
                resource: "https://docs.python.org/3/tutorial/introduction.html"
                sort_order: 1
                created_at: "2026-06-08T12:00:00Z"
                completed_at: null
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
  /api/tasks/{tid}/start:
    patch:
      summary: Start a task (pending → doing)
      description: |
        Transitions a task from 'pending' to 'doing'. Fails with 409 if
        the task has unmet dependencies or is not in 'pending' status.
      operationId: startTask
      tags: [Tasks]
      parameters:
        - name: tid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatusResponse"
              example:
                id: "t1"
                status: "doing"
                previous_status: "pending"
                started_at: "2026-06-08T14:30:00Z"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "409":
          description: Invalid state transition or dependency blocked
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/InvalidTransitionError"
                  - $ref: "#/components/schemas/DependencyBlockedError"
              examples:
                invalid_state:
                  value:
                    error: "INVALID_TRANSITION"
                    message: "Task is already in 'doing' status"
                    current_status: "doing"
                    required_status: "pending"
                dependency_blocked:
                  value:
                    error: "DEPENDENCY_BLOCKED"
                    message: "Cannot start task — dependencies not met"
                    blocked_by: ["t1"]
  /api/tasks/{tid}/checkin:
    post:
      summary: Complete a task (doing → done)
      description: |
        Marks a task as 'done', awards XP, and optionally progresses the
        linked project's dream reward tracker.
      operationId: checkinTask
      tags: [Tasks]
      parameters:
        - name: tid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Check-in successful with XP reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResponse"
              example:
                id: "t1"
                status: "done"
                previous_status: "doing"
                completed_at: "2026-06-08T15:00:00Z"
                xp_earned: 20
                dream_value_earned: 2
                new_total_xp: 350
                new_level: 4
                xp_to_next_level: 50
                streak: 7
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "409":
          description: Task is not in 'doing' status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidTransitionError"
              example:
                error: "INVALID_TRANSITION"
                message: "Can only check in tasks that are in 'doing' status"
                current_status: "pending"
                required_status: "doing"

  # ──────────────────────────────────────────────
  #  PROGRESS & REWARDS
  # ──────────────────────────────────────────────
...
components:
  schemas:
...
    TaskSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [theory, practice, output]
        xp:
          type: integer
        status:
          type: string
          enum: [pending, doing, done]
        sort_order:
          type: integer
    TaskDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        stage_title:
          type: string
        project_id:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [theory, practice, output]
        xp:
          type: integer
        estimate:
          type: integer
          description: Estimated time in minutes
        status:
          type: string
          enum: [pending, doing, done]
        depends:
          type: array
          items:
            type: string
          description: IDs of tasks that must be completed first
        check:
          type: string
          nullable: true
          description: Acceptance criteria
        resource:
          type: string
          nullable: true
          description: Learning resource URL or reference
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    TodayTask:
      allOf:
        - $ref: "#/components/schemas/TaskSummary"
        - type: object
          properties:
            stage_title:
              type: string
            depends:
              type: array
              items:
                type: string
            blocked:
              type: boolean
              description: True if dependencies are unmet
            check:
              type: string
              nullable: true
    TaskStatusResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, doing, done]
        previous_status:
          type: string
          enum: [pending, doing, done]
        started_at:
          type: string
          format: date-time
    CheckinResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [done]
        previous_status:
          type: string
          enum: [doing]
        completed_at:
          type: string
          format: date-time
        xp_earned:
          type: integer
        dream_value_earned:
          type: integer
        new_total_xp:
          type: integer
        new_level:
          type: integer
        xp_to_next_level:
          type: integer
        streak:
          type: integer
...
    NotFoundError:
      type: object
      properties:
        detail:
          type: string
      example:
        detail: "Project not found"
...
    InvalidTransitionError:
      type: object
      properties:
        error:
          type: string
          enum: [INVALID_TRANSITION]
        message:
          type: string
        current_status:
          type: string
        required_status:
          type: string
    DependencyBlockedError:
      type: object
      properties:
        error:
          type: string
          enum: [DEPENDENCY_BLOCKED]
        message:
          type: string
        blocked_by:
          type: array
          items:
            type: string